refactor: worker to db trigger
